authotize user for likes and make view for all post by user and total likes

+++ app/Http/Controllers/PostController.php
@@ -29,6 +29,9 @@ class PostController extends Controller
     }
     public function delete(Post $post)
     {
+        //Authorise post user
+        $this->authorize('delete', $post);
+
         $post->delete();
         return back();
     }
